bg_api/requirements.txt
Flask==2.2.3
PyJWT==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.10
Werkzeug==2.2.3
psycopg2==2.9.6
bcrypt==4.0.1